Cleanup + remove dead code
