+++ PlayCover/Views/App Views/PlayAppView.swift
@@ -175,6 +175,26 @@ struct PlayAppView: View {
             print("Error when attempting to remove Twitter session cookie: \(error)")
         }
     }
+
+    func deletePreferences(app: String) {
+        let plistURL = FileManager.default.homeDirectoryForCurrentUser
+            .appendingPathComponent("Library")
+            .appendingPathComponent("Containers")
+            .appendingEscapedPathComponent(app)
+            .appendingPathComponent("Data")
+            .appendingPathComponent("Library")
+            .appendingPathComponent("Preferences")
+            .appendingEscapedPathComponent(app)
+            .appendingPathExtension("plist")
+
+        guard FileManager.default.fileExists(atPath: plistURL.path) else { return }
+
+        do {
+            try FileManager.default.removeItem(atPath: plistURL.path)
+        } catch {
+            Log.shared.log("\(error)", isError: true)
+        }
+    }
 }
 
 struct PlayAppConditionalView: View {
